Remove our usage of ssize_t

+++ Core/HLE/sceIo.cpp
@@ -967,8 +967,8 @@ static u32 npdrmRead(FileNode *f, u8 *data, int size) {
 	block  = pgd->file_offset/pgd->block_size;
 	offset = pgd->file_offset%pgd->block_size;
 
-	if (size > pgd->data_size)
-		size = pgd->data_size;
+	if (size > (int)pgd->data_size)
+		size = (int)pgd->data_size;
 	remain_size = size;
 
 	while(remain_size){
